provide runtests.py for running tests in CI
[#170913126]
