+++ .gitignore
@@ -8,3 +8,4 @@ build
 node_modules/
 static/gen/
 yarn-error.log
+TEST*.xml
