+++ models/__init__.py
@@ -1,6 +1,3 @@
-from django.contrib.auth.models import User
-from django.db import models
-
 from tracker.models.bid import Bid, BidSuggestion, DonationBid
 from tracker.models.country import Country, CountryRegion
 from tracker.models.donation import Donation, Donor, DonorCache
@@ -11,6 +8,7 @@ from tracker.models.event import (
     SpeedRun,
     Submission,
 )
+from tracker.models.log import Log
 from tracker.models.mod_filter import AmountFilter, WordFilter
 from tracker.models.prize import (
     DonorPrizeEntry,
@@ -19,6 +17,7 @@ from tracker.models.prize import (
     PrizeKey,
     PrizeWinner,
 )
+from tracker.models.profile import UserProfile
 
 __all__ = [
     'Event',
@@ -37,59 +36,10 @@ __all__ = [
     'SpeedRun',
     'Runner',
     'Submission',
-    'UserProfile',
-    'Log',
     'Country',
     'CountryRegion',
     'WordFilter',
     'AmountFilter',
+    'Log',
+    'UserProfile',
 ]
-
-
-class UserProfile(models.Model):
-    user = models.OneToOneField(User)
-    prepend = models.CharField('Template Prepend', max_length=64, blank=True)
-
-    class Meta:
-        app_label = 'tracker'
-        verbose_name = 'User Profile'
-        permissions = (
-            ('show_rendertime', 'Can view page render times'),
-            ('show_queries', 'Can view database queries'),
-            ('can_search', 'Can use search url'),
-        )
-
-    def __str__(self):
-        return str(self.user)
-
-
-class Log(models.Model):
-    timestamp = models.DateTimeField(auto_now_add=True, verbose_name='Timestamp')
-    category = models.CharField(
-        max_length=64, default='other', blank=False, null=False, verbose_name='Category'
-    )
-    message = models.TextField(blank=True, null=False, verbose_name='Message')
-    event = models.ForeignKey('Event', blank=True, null=True, on_delete=models.PROTECT)
-    user = models.ForeignKey(User, blank=True, null=True)
-
-    class Meta:
-        app_label = 'tracker'
-        verbose_name = 'Log'
-        permissions = (
-            # TODO: neither of these really make sense
-            ('can_view_log', 'Can view tracker logs'),
-            ('can_change_log', 'Can change tracker logs'),
-        )
-        ordering = ['-timestamp']
-
-    def __str__(self):
-        result = str(self.timestamp)
-        if self.event:
-            result += ' (' + self.event.short + ')'
-        result += ' -- ' + self.category
-        if self.message:
-            m = self.message
-            if len(m) > 18:
-                m = m[:15] + '...'
-            result += ': ' + m
-        return result
