+++ tests/test_donate.py
@@ -5,7 +5,7 @@ from django.urls import reverse
 
 import tracker.forms as forms
 import tracker.models as models
-from . import today_noon, tomorrow_noon, long_ago_noon
+from .util import today_noon, tomorrow_noon, long_ago_noon
 
 
 class TestDonorNameAssignment(TransactionTestCase):
